Added if statement to better handle missing files.

+++ docker/config/volume-files/server-startup.sh
@@ -4,20 +4,25 @@ set -e
 
 if [ "$1" = 'control' ]
 then
-    cp /root/files/private_keys/ansible-docker-lab /root/.ssh/id_rsa
-    chmod 600 /root/.ssh/id_rsa
+    if [[ -e "/root/files/private_keys/ansible-docker-lab" ]]; then
+        cp /root/files/private_keys/ansible-docker-lab /root/.ssh/id_rsa
+        chmod 600 /root/.ssh/id_rsa
+    fi
     
-    cp /root/files/private_keys/ansible-docker-lab /home/demo/.ssh/id_rsa
-    chown demo:demo /home/demo/.ssh/id_rsa
-    chmod 600 /home/demo/.ssh/id_rsa
+    if [[ -e "/root/files/private_keys/ansible-docker-lab" ]]; then
+        cp /root/files/private_keys/ansible-docker-lab /home/demo/.ssh/id_rsa
+        chown demo:demo /home/demo/.ssh/id_rsa
+        chmod 600 /home/demo/.ssh/id_rsa
+    fi
 else
-    cp /root/files/authorized_keys.pub /root/.ssh/authorized_keys
-    chmod 600 /root/.ssh/authorized_keys
+    if [[ -e "/root/files/authorized_keys.pub" ]]; then
+        cp /root/files/authorized_keys.pub /root/.ssh/authorized_keys
+        chmod 600 /root/.ssh/authorized_keys
     
-    cp /root/files/authorized_keys.pub /home/demo/.ssh/authorized_keys
-    chown demo:demo /home/demo/.ssh/authorized_keys
-    chmod 600 /home/demo/.ssh/authorized_keys
-    
+        cp /root/files/authorized_keys.pub /home/demo/.ssh/authorized_keys
+        chown demo:demo /home/demo/.ssh/authorized_keys
+        chmod 600 /home/demo/.ssh/authorized_keys
+    fi
 fi
 
 exec /usr/sbin/sshd -D
